Tear down PermissionError bug fix for Windows OS
