+++ tests/views/test_receipts.py
@@ -1,10 +1,8 @@
-import uuid
-
 from django.contrib.messages import get_messages
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.urls import reverse, resolve
 
-from backend.models import Receipt, ReceiptDownloadToken
+from backend.models import Receipt
 from tests.handler import ViewTestCase
 
 
@@ -84,45 +82,3 @@ class ReceiptsAPITestCase(ViewTestCase):
         messages = list(get_messages(response.wsgi_request))
         self.assertEqual(len(messages), 1)
         self.assertEqual(str(messages[0]), "No image found")
-
-
-class ReceiptDownloadEndpointsTest(ViewTestCase):
-    def setUp(self):
-        super().setUp()
-        self.receipt = Receipt.objects.create(
-            user=self.log_in_user, image=SimpleUploadedFile("mock_image.jpg", b"image_content", "image/jpeg")
-        )
-        self.token = ReceiptDownloadToken.objects.create(user=self.log_in_user, file=self.receipt)
-        self.download_receipt_url = reverse("api:receipts:download_receipt", args=[self.token.token])
-        self.generate_download_link_url = reverse("api:receipts:generate_download_link", args=[self.receipt.id])
-
-    def test_download_receipt_valid_token(self):
-        self.login_user()
-        response = self.client.get(self.download_receipt_url)
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response["Content-Type"], "image/jpeg")
-
-    def test_download_receipt_invalid_token(self):
-        self.login_user()
-        invalid_token = uuid.uuid4()  # Generate a valid UUID
-        invalid_url = reverse("api:receipts:download_receipt", args=[invalid_token])
-        response = self.client.get(invalid_url)
-        self.assertEqual(response.status_code, 404)  # Update expected status code
-
-    def test_download_receipt_used_token(self):
-        self.login_user()
-        self.client.get(self.download_receipt_url)  # Use the token once
-        response = self.client.get(self.download_receipt_url)  # Try to use the token again
-        self.assertEqual(response.status_code, 404)  # Expect a 404 response
-
-    def test_generate_download_link_valid_receipt(self):
-        self.login_user()
-        response = self.client.get(self.generate_download_link_url)
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response["Content-Type"], "application/json")
-
-    def test_generate_download_link_invalid_receipt(self):
-        self.login_user()
-        invalid_url = reverse("api:receipts:generate_download_link", args=[9999])  # Assuming 9999 is an invalid receipt id
-        response = self.client.get(invalid_url)
-        self.assertEqual(response.status_code, 404)
